Add `length: never` field to ArrayLike
Consider the following case:
```ts
type ArrayMembers<A> = A[Exclude<keyof A, keyof Array<any> | "length">];
type B = ArrayMembers<[1, 2, 4]>;
```

In order to properly get all the keys in a tuple of its indeces, one must `Exclude` from `keyof A` `keyof Array<any> | "length"`. This also creates problem in more complicated types where a given type is proven to extend Array but TS still can't validate that it has the `length` member (which in our case, would only exist in the type of fixed-length arrays). In order to circumvent this issue, one must check `keyof T extends "length" ? T["length"]`, however, this only works in TS 3.9 and above and I think only adds unnecessarily to the learning curve for roblox-ts. Hence, I think it is best to add a `length: never` property to the Array type.