Add DS V2 Types (#490)
* Add DS V2 Types

* Pretty definitions

* Move definitions to DataStore from GlobalDataStore

Previously many of the new changes were applied under the
GlobalDataStore interface. This is incorrect, as Roblox will error when
attempting to use some of the signatures for OrderedDataStore calls.
Abstracting the methods to a new DataStore type is the correct move.

* Fix GetAsync signature for GlobalDataStore

"OrderedDataStore does not support versioning and metadata, so
DataStoreKeyInfo will always be nil for keys in an OrderedDataStore."

Moving the DataStoreKeyInfo signature to the DataStore type corrects
this issue.

* Fix definitions

Fixes many definitions and splits the types up more.

* Update DataStoreVersionPages type

Correctly marks the type of GetCurrentPage.

Extending the Pages type will not work, as it gets a type mismatch.